Replace font size with typography
